# Make broadcast eltype widening type-stable