The broadcast code that copies to the new array was type-unstable, causing allocation and slowdown in `v2`:

Before:

```julia
julia> using BenchmarkTools

julia> v = fill(2.0, 50000);

julia> v2 = vcat(v, missing);

julia> v3 = vcat(missing, v);

julia> @btime $v .* 3.0;
  49.728 μs (2 allocations: 390.70 KiB)

julia> @btime $v2 .* 3.0;
  1.490 ms (99498 allocations: 2.33 MiB)

julia> @btime $v3 .* 3.0;
  103.577 μs (7 allocations: 439.77 KiB)
```

After:

```julia
julia> @btime $v .* 3.0;
  50.147 μs (2 allocations: 390.70 KiB)

julia> @btime $v2 .* 3.0;
  147.158 μs (8 allocations: 830.45 KiB)

julia> @btime $v3 .* 3.0;
  92.611 μs (7 allocations: 439.77 KiB)
```

This fixes the second half of https://github.com/JuliaLang/julia/issues/30455. First half was fixed by https://github.com/JuliaLang/julia/pull/30480