@mbauman 2019-02-25
Nice.  Makes sense.  I know it's a small cost, but since we're outlining here, perhaps we should just outline everything after that `similar` call.  E.g.,

```julia
newdest = Base.similar(dest, promote_typejoin(T, typeof(val)))
return restart_copyto_nonleaf!(newdest, dest, bc, val, I, iter, state, count)
```

```julia
function restart_copyto_nonleaf!(newdest, dest, bc, val, I, iter, state, count)
    # Function barrier that makes the copying to newdest type stable
    for II in Iterators.take(iter, count)
        newdest[II] = dest[II]
    end
    newdest[I] = val
    return copyto_nonleaf!(newdest, bc, iter, state, count+1)
end
```

That should just have one dynamic dispatch instead of three.  In my quick tests of the cases you listed, it looks like it saves another ~25%.