@mbauman 2019-02-27
The Travis failures look unrelated — 64 bit linux is an assertion failure in the Profile tests and Mac is a Pkg timeout.