@mbauman 2019-02-27
We're trying to be a bit more conservative in backporting performance fixes, but this is about as straightforward as a patch could be — it's literally just copy-pasting code into an outlined function.  I'd support a backport.